Refactored to put pipeline stuff into data_management
Added a clean_dataframe method to handle non-sklearn label replacement stuff.

Added a debug mode to export the training data to CSV when needed

Current status:
-Getting a value error when running the models (run_models.py). Unclear what this is from. Should try dropping columns one by one until we identify which column this is coming from. Big change since last time on this was adding in the rent column, but I think that imputation is currently handling all the values in that appropriately. Maybe round off the imputed value? seems odd source of problem....
